+++ src/edu/uclouvain/core/nodus/compute/costs/CostParser.java
@@ -752,7 +752,6 @@ public class CostParser {
          * The volume is a number of passenger car units (PCU's)
          */
         setVariable(NodusC.VARNAME_VOLUME, rl.getCurrentPassengerCarUnits(vl));
-        System.err.println("VOLUME on " + vl.getBeginVirtualNode().getRealLinkId() + " =  " + rl.getCurrentPassengerCarUnits(vl));
 
       } else {
         fieldName = nodeFieldName[layerIndex];